Add clear completed
